requirements.txt
Django==3.0
line-bot-sdk==1.15.0
gunicorn==20.0.4
